fix(winemetal): make DispatchData pointer cast portable

+++ src/winemetal/winemetal_thunks.c
@@ -1005,7 +1005,7 @@ WINEMETAL_API uint64_t
 DispatchData_copy(obj_handle_t data, void *destination, uint64_t capacity) {
   struct unixcall_dispatchdata_copy params;
   params.data = data;
-  params.destination = (uint64_t)destination;
+  params.destination = (uint64_t)(uintptr_t)destination;
   params.capacity = capacity;
   params.ret_size = 0;
   UNIX_CALL(unix_dispatchdata_copy, &params);
